SimultaneousState saving vec conversion bug
SimultaneousState.save_list was not working correctly due to a conversion order of operations issue.
The conversion between python and rust for list of SimultaneousStates was unintentionally dropping the FOV information.
This was leading to errors when attempting to save simulation results which included this information.

This was fixed by changing the order of operations to try SimState conversions before list conversions.
Error messages were edited to increase the clarity of the bug.