Prep for final WISE release (#112)
* Prep for final WISE release

* removing time scaling from Covariance epoch
Covariances can have multiple parameters of time, it seems better to
leave time in UTC jd consistently across all cov values. The user will
have to make the conversion. This should be only a small error if
forgotten, however it would be a much larger error if only the epoch
is converted.

* fix example plotting